Move remaining implementations of "from_list" to its module
